+++ pages/data/study.tsx
@@ -1,14 +1,22 @@
 import React from "react";
 import PageTitle from "../../src/components/common/Layout/PageTitle";
 import Banner from "../../src/components/common/Layout/Banner";
-import StudyMaterialTable from "../../src/components/data/StudyMaterialTable";
+import MaterialTable from "../../src/components/data/MaterialTable";
 
 function Research() {
     return (
         <div>
-            <StudyMaterialTable material='학습 자료' />
+            <MaterialTable material="학습 자료" />
         </div>
     );
 }
 
 export default Research;
+
+/* 학습자료
+    create ✓
+    read   ✓
+    update ✓
+    delete ✓
+    파일 수정관련해서 수정시(no file change)? 일단 현재파일로 설정
+*/
