+++ src/components/data/MaterialTableButton.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from "react";
-import { StudyMaterialTableContext } from "./StudyMaterialTable";
-import { StudyMaterial } from "../../../typing";
+import { MaterialTableContext } from "./MaterialTable";
+import { StudyMaterial, PaperMaterial } from "../../../typing";
 
 function MaterialTableAddButton() {
     const { setIsEditing, isEditing, setSelectedMaterial } =
-        useContext(StudyMaterialTableContext);
+        useContext(MaterialTableContext);
 
     return (
         <button
@@ -19,12 +19,12 @@ function MaterialTableAddButton() {
     );
 }
 
-function MaterialTableUpdateButton({
+function MaterialTableUpdateButton<M extends StudyMaterial | PaperMaterial>({
     data,
     docID,
     index,
 }: {
-    data: StudyMaterial;
+    data: M;
     docID: string;
     index: number;
 }) {
@@ -33,7 +33,7 @@ function MaterialTableUpdateButton({
         setSelectedMaterial,
         setSelectedDocId,
         setSelectedIndex,
-    } = useContext(StudyMaterialTableContext);
+    } = useContext(MaterialTableContext);
 
     return (
         <button
@@ -54,13 +54,13 @@ function MaterialTableUpdateButton({
 }
 
 function MaterialTableDeleteButton({ docID }: { docID: string }) {
-    const { deleteDocument } = useContext(StudyMaterialTableContext);
+    const { deleteDocument } = useContext(MaterialTableContext);
     return <button onClick={() => deleteDocument(docID)}>삭제</button>;
 }
 
 function MaterialTableCancelButton() {
     const { setIsEditing, setSelectedMaterial, setSelectedIndex } =
-        useContext(StudyMaterialTableContext);
+        useContext(MaterialTableContext);
     return (
         <button
             className='border p-1'
